rgw/s3vectors: prevent cascade delete of indexes on bucket delete
return 409 according to AWS spec.

+++ src/rgw/rgw_rest_s3vector.cc
@@ -420,12 +420,14 @@ private:
       ldpp_dout(this, 1) << "ERROR: failed to load s3vector bucket " << bucket_id << ". error: " << op_ret << dendl;
       return;
     }
+    op_ret = rgw::s3vector::delete_vector_bucket(configuration, driver, s->user.get(), &s->bucket_tenant, this, y);
+    if (op_ret < 0) {
+      return;
+    }
     op_ret = bucket->remove(this, false, y);
     if (op_ret < 0) {
       ldpp_dout(this, 1) << "ERROR: failed to delete s3vector bucket " << bucket_id << ". error: " << op_ret << dendl;
-      return;
     }
-    op_ret = rgw::s3vector::delete_vector_bucket(configuration, driver, s->user.get(), &s->bucket_tenant, this, y);
   }
 };
